sanyun-back-end/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cx-Oracle==8.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7 